SVHN
This project builds on the work of publicly available GitHub code . We used this as a foundation in developing a unified approach that integrates localization, segmentation, and recognition through the use of a deep convolutional network as proposed by Goodfellow. Additional code was added, particularly relating to extending the neural network, evaluation and creation of performance metrics.